+++ lib/src/index.ts
@@ -1,2 +1,0 @@
-export { bewegung } from "./bewegung";
-export { Bewegung } from "./types";