+++ lib/src/utils/constants.ts
@@ -22,6 +22,3 @@ export const enum Attributes {
 	cssReset = "data-bewegungs-css-reset",
 	replace = "data-bewegungs-replace",
 }
-
-//TODO:remove later if debugging is not needed anymore
-export const sleep = (duration: number) => new Promise((resolve) => setTimeout(resolve, duration));