+++ lib/CHANGELOG.md
@@ -1,4 +1,4 @@
-#1
+#latest:
 
-- filtered blocks with faulty root element
-- added a queueing system to record user inputs while reading the dom
+- use checkpoint animations to time the calculations of the actual animations
+- together with the usage of the timekeeper startTime reduces a lot of complexity for orchestrating the animations playstate and progress